#2 jpa 설정 변경 및 Long type으로 변경

+++ sks-api/src/main/java/com/sks/api/domain/PostsRepository.java
@@ -2,6 +2,6 @@ package com.sks.api.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PostsRepository extends CrudRepository<Posts, Integer> {
+public interface PostsRepository extends CrudRepository<Posts, Long> {
 
 }
